+++ modules/backendscheduler/provider/metrics.go
@@ -16,4 +16,14 @@ var (
 		Name:      "compaction_tenant_reset_total",
 		Help:      "The number of times the tenant is changed",
 	}, []string{"tenant"})
+	metricTenantBackoff = promauto.NewCounter(prometheus.CounterOpts{
+		Namespace: "tempo_backend_scheduler",
+		Name:      "compaction_tenant_backoff_total",
+		Help:      "The number of times the backoff is triggered",
+	})
+	metricTenantEmptyJob = promauto.NewCounter(prometheus.CounterOpts{
+		Namespace: "tempo_backend_scheduler",
+		Name:      "compaction_tenant_empty_job_total",
+		Help:      "The number of times an empty job was received from the priority queue",
+	})
 )